refactor(e2e): extract the failure taxonomy and navigation helpers
The three failure classes are one concept documented as one taxonomy in
AGENTS.md, but were implemented three times across two files, which left
seed.js importing SetupError from extension.js. They now live together in
errors.js, so adding or changing a category is one edit in one place.

gotoOrPrecondition moves out of cookies.js, which had become the home for
cookie assertions, the error classes and navigation alike. It joins a new
openSite helper that replaces the newPage-then-navigate pair repeated seven
times across the scenarios.

Also drops a keeperPage variable that was assigned and never read, and two
exports from extension.js with no consumer.
